+++ plan.php
@@ -743,6 +743,9 @@ function ip()
     return assert\validate('validate_ip');
 }
 
+/**
+ * @deprecated
+ */
 function regexp()
 {
     return assert\validate('validate_regexp');